+++ context/app/api/mock_client.py
@@ -2,7 +2,7 @@ import json
 
 from datauri import DataURI
 
-from .vitessce_confs.base_confs import ConfCells
+from .vitessce_confs.base_builders import ConfCells
 from .client import ApiClient
