+++ context/app/api/vitessce_confs/sprm_builders.py
@@ -1,5 +1,3 @@
-import urllib
-from pathlib import Path
 import re
 from collections import namedtuple
 
@@ -14,220 +12,15 @@ from vitessce import (
     FileType as ft,
 )
 
+from .base_builders import ConfCells, ViewConfBuilder
 from .utils import get_matches
-from .paths import SPRM_JSON_DIR, IMAGE_PYRAMID_DIR, OFFSETS_DIR
-
-MOCK_URL = "https://example.com"
-
-
-ConfCells = namedtuple('ConfCells', ['conf', 'cells'])
+from .paths import SPRM_JSON_DIR, STITCHED_REGEX, CODEX_TILE_DIR, TILE_REGEX, STITCHED_IMAGE_DIR
+from .imaging_builders import ImagePyramidViewConfBuilder
 
 
-class NullViewConfBuilder():
-
-    def get_conf_cells(self):
-        return ConfCells(None, None)
-
-
-class ViewConfBuilder:
-    def __init__(self, entity=None, groups_token=None, is_mock=False):
-        """Object for building the vitessce configuration.
-        :param dict entity: Entity response from search index (from the entity API)
-        :param str groups_token: Groups token for use in authenticating API
-        :param bool is_mock: Wether or not this class is being mocked.
-
-        >>> vc = ViewConfBuilder(
-        ...     entity={"uuid": "uuid"}, groups_token='groups_token', is_mock=True)
-
-        """
-
-        self._uuid = entity["uuid"]
-        self._groups_token = groups_token
-        self._entity = entity
-        self._is_mock = is_mock
-        self._files = []
-
-    def get_conf_cells(self):
-        raise NotImplementedError
-
-    def _replace_url_in_file(self, file):
-        """Replace url in incoming file object
-        :param dict file: File dict which will have its rel_path replaced by url
-        :rtype: dict The file with rel_path replaced by url
-        >>> from pprint import pprint
-        >>> vc = ViewConfBuilder(
-        ...     entity={"uuid": "uuid"}, groups_token='groups_token', is_mock=True)
-        >>> file = { 'data_type': 'CELLS', 'file_type': 'cells.json', 'rel_path': 'cells.json' }
-        >>> pprint(vc._replace_url_in_file(file))
-        {'data_type': 'CELLS',\n\
-         'file_type': 'cells.json',\n\
-         'url': 'https://example.com/uuid/cells.json?token=groups_token'}
-        """
-
-        return {
-            "data_type": file["data_type"],
-            "file_type": file["file_type"],
-            "url": self._build_assets_url(file["rel_path"]),
-        }
-
-    def _build_assets_url(self, rel_path, use_token=True):
-        """Create a url for an asset.
-        :param str rel_path: The path off of which the url should be built
-        :param bool use_token: Whether or not to append a groups token to the URL, default True
-        :rtype: dict The file with rel_path replaced by url
-
-        >>> vc = ViewConfBuilder(
-        ...     entity={"uuid": "uuid"}, groups_token='groups_token', is_mock=True)
-        >>> vc._build_assets_url("rel_path/to/clusters.ome.tiff")
-        'https://example.com/uuid/rel_path/to/clusters.ome.tiff?token=groups_token'
-
-        """
-        if not self._is_mock:
-            assets_endpoint = current_app.config["ASSETS_ENDPOINT"]
-        else:
-            assets_endpoint = MOCK_URL
-        base_url = urllib.parse.urljoin(assets_endpoint, f"{self._uuid}/{rel_path}")
-        token_param = urllib.parse.urlencode({"token": self._groups_token})
-        return f"{base_url}?{token_param}" if use_token else base_url
-
-    def _get_request_init(self):
-        """Get request headers for requestInit parameter in Vitessce conf.
-        This is needed for non-public zarr stores because the client forms URLs for zarr chunks,
-        not the above _build_assets_url function.
-
-        >>> entity = {"uuid": "uuid", "status": "QA"}
-        >>> vc = ViewConfBuilder(entity=entity, groups_token='groups_token', is_mock=True)
-        >>> vc._get_request_init()
-        {'headers': {'Authorization': 'Bearer groups_token'}}
-        >>> entity = {"uuid": "uuid", "status": "Published"}
-        >>> vc = ViewConfBuilder(entity=entity, groups_token='groups_token', is_mock=True)
-        >>> assert vc._get_request_init() is None # None because dataset is Published (public)
-        """
-        request_init = {"headers": {"Authorization": f"Bearer {self._groups_token}"}}
-        # Extra headers outside of a select few cause extra CORS-preflight requests which
-        # can slow down the webpage.  If the dataset is published, we don't need to use
-        # heaeder to authenticate access to the assets API.
-        # See: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS#simple_requests
-        use_request_init = False if self._entity["status"] == "Published" else True
-        return request_init if use_request_init else None
-
-    def _get_file_paths(self):
-        """Get all rel_path keys from the entity dict.
-
-        >>> files = [{ "rel_path": "path/to/file" }, { "rel_path": "path/to/other_file" }]
-        >>> entity = {"uuid": "uuid", "files": files}
-        >>> vc = ViewConfBuilder(entity=entity, groups_token='groups_token', is_mock=True)
-        >>> vc._get_file_paths()
-        ['path/to/file', 'path/to/other_file']
-        """
-        return [file["rel_path"] for file in self._entity["files"]]
-
-
-class AbstractImagingViewConfBuilder(ViewConfBuilder):
-    def _get_img_and_offset_url(self, img_path, img_dir):
-        """Create a url for the offsets and img.
-        :param str img_path: The path of the image
-        :param str img_dir: The image-specific part of the path to be
-        replaced by the OFFSETS_DIR constant.
-        :rtype: tuple The image url and the offsets url
-
-        >>> from pprint import pprint
-        >>> vc = AbstractImagingViewConfBuilder(entity={ "uuid": "uuid" },\
-            groups_token='groups_token', is_mock=True)
-        >>> pprint(vc._get_img_and_offset_url("rel_path/to/clusters.ome.tiff",\
-            "rel_path/to"))
-        ('https://example.com/uuid/rel_path/to/clusters.ome.tiff?token=groups_token',\n\
-         'https://example.com/uuid/output_offsets/clusters.offsets.json?token=groups_token')
-
-        """
-        img_url = self._build_assets_url(img_path)
-        return (
-            img_url,
-            str(
-                re.sub(
-                    r"ome\.tiff?",
-                    "offsets.json",
-                    re.sub(img_dir, OFFSETS_DIR, img_url),
-                )
-            ),
-        )
-
-    def _setup_view_config_raster(self, vc, dataset, disable_3d=[]):
-        vc.add_view(dataset, cm.SPATIAL, x=3, y=0, w=9, h=12)
-        vc.add_view(dataset, cm.DESCRIPTION, x=0, y=8, w=3, h=4)
-        vc.add_view(dataset, cm.LAYER_CONTROLLER, x=0, y=0, w=3, h=8).set_props(
-            disable3d=disable_3d
-        )
-        return vc
-
-
-class ImagePyramidViewConfBuilder(AbstractImagingViewConfBuilder):
-    def __init__(self, entity, groups_token, is_mock=False):
-        """Wrapper class for creating a standard view configuration for image pyramids,
-        i.e for high resolution viz-lifted imaging datasets like
-        https://portal.hubmapconsortium.org/browse/dataset/dc289471333309925e46ceb9bafafaf4
-        """
-        self.image_pyramid_regex = IMAGE_PYRAMID_DIR
-        super().__init__(entity, groups_token, is_mock)
-
-    def get_conf_cells(self):
-        file_paths_found = self._get_file_paths()
-        found_images = get_matches(
-            file_paths_found, self.image_pyramid_regex + r".*\.ome\.tiff?$",
-        )
-        if len(found_images) == 0:
-            message = (
-                f"Image pyramid assay with uuid {self._uuid} has no matching files"
-            )
-            raise FileNotFoundError(message)
-
-        vc = VitessceConfig(name="HuBMAP Data Portal")
-        dataset = vc.add_dataset(name="Visualization Files")
-        images = []
-        for img_path in found_images:
-            img_url, offsets_url = self._get_img_and_offset_url(
-                img_path, self.image_pyramid_regex
-            )
-            images.append(
-                OmeTiffWrapper(
-                    img_url=img_url, offsets_url=offsets_url, name=Path(img_path).name
-                )
-            )
-        dataset = dataset.add_object(MultiImageWrapper(images))
-        vc = self._setup_view_config_raster(vc, dataset)
-        conf = vc.to_dict()
-        # Don't want to render all layers
-        del conf["datasets"][0]["files"][0]["options"]["renderLayers"]
-        return ConfCells(conf, None)
-
-
-class AbstractScatterplotViewConfBuilder(ViewConfBuilder):
-    """Base class for subclasses creating a JSON-backed scatterplot for
-    "first generation" RNA-seq and ATAC-seq data like
-    https://portal.hubmapconsortium.org/browse/dataset/d4493657cde29702c5ed73932da5317c
-    from h5ad-to-arrow.cwl.
-    """
-
-    def get_conf_cells(self):
-        file_paths_expected = [file["rel_path"] for file in self._files]
-        file_paths_found = self._get_file_paths()
-        # We need to check that the files we expect actually exist.
-        # This is due to the volatility of the datasets.
-        if not set(file_paths_expected).issubset(set(file_paths_found)):
-            message = f'Files for uuid "{self._uuid}" not found as expected.'
-            raise FileNotFoundError(message)
-        vc = VitessceConfig(name="HuBMAP Data Portal")
-        dataset = vc.add_dataset(name="Visualization Files")
-        # The sublcass initializes _files in its __init__ method
-        for file in self._files:
-            dataset = dataset.add_file(**(self._replace_url_in_file(file)))
-        vc = self._setup_scatterplot_view_config(vc, dataset)
-        return ConfCells(vc.to_dict(), None)
-
-    def _setup_scatterplot_view_config(self, vc, dataset):
-        vc.add_view(dataset, cm.SCATTERPLOT, mapping="UMAP", x=0, y=0, w=9, h=12)
-        vc.add_view(dataset, cm.CELL_SETS, x=9, y=0, w=3, h=12)
-        return vc
+class CytokitSPRMViewConfigError(Exception):
+    """Raised when one of the individual SPRM view configs errors out"""
+    pass
 
 
 class SPRMViewConfBuilder(ImagePyramidViewConfBuilder):
@@ -428,3 +221,73 @@ class SPRMAnnDataViewConfBuilder(SPRMViewConfBuilder):
             variablesLabelOverride="antigen", transpose=True
         )
         return vc
+
+
+class StitchedCytokitSPRMViewConfBuilder(ViewConfBuilder):
+    """Wrapper class for generating multiple "second generation" stitched AnnData-backed SPRM
+    Vitessce configurations via SPRMAnnDataViewConfBuilder,
+    used for datasets with multiple regions.
+    These are from post-August 2020 Cytokit datasets (stitched).
+    """
+
+    def get_conf_cells(self):
+        file_paths_found = [file["rel_path"] for file in self._entity["files"]]
+        found_regions = get_matches(file_paths_found, STITCHED_REGEX)
+        if len(found_regions) == 0:
+            raise FileNotFoundError(
+                f"Cytokit SPRM assay with uuid {self._uuid} has no matching regions; "
+                f"No file matches for '{STITCHED_REGEX}'."
+            )
+        confs = []
+        for region in sorted(found_regions):
+            vc = SPRMAnnDataViewConfBuilder(
+                entity=self._entity,
+                groups_token=self._groups_token,
+                is_mock=self._is_mock,
+                base_name=region,
+                imaging_path=STITCHED_IMAGE_DIR,
+                mask_path=STITCHED_IMAGE_DIR.replace('expressions', 'mask'),
+                image_name=f"{region}_stitched_expressions",
+                mask_name=f"{region}_stitched_mask"
+            )
+            conf = vc.get_conf_cells().conf
+            if conf == {}:
+                raise CytokitSPRMViewConfigError(
+                    f"Cytokit SPRM assay with uuid {self._uuid} has empty view\
+                        config for region '{region}'"
+                )
+            confs.append(conf)
+        return ConfCells(confs if len(confs) > 1 else confs[0], None)
+
+
+class TiledSPRMViewConfBuilder(ViewConfBuilder):
+    """Wrapper class for generating many "first generation"
+    non-stitched JSON-backed SPRM Vitessce configurations,
+    one per tile per region, via SPRMJSONViewConfBuilder.
+    """
+
+    def get_conf_cells(self):
+        file_paths_found = [file["rel_path"] for file in self._entity["files"]]
+        found_tiles = get_matches(
+            file_paths_found,
+            TILE_REGEX) or get_matches(
+            file_paths_found,
+            STITCHED_REGEX)
+        if len(found_tiles) == 0:
+            message = f'Cytokit SPRM assay with uuid {self._uuid} has no matching tiles'
+            raise FileNotFoundError(message)
+        confs = []
+        for tile in sorted(found_tiles):
+            vc = SPRMJSONViewConfBuilder(
+                entity=self._entity,
+                groups_token=self._groups_token,
+                is_mock=self._is_mock,
+                base_name=tile,
+                imaging_path=CODEX_TILE_DIR
+            )
+            conf = vc.get_conf_cells().conf
+            if conf == {}:
+                message = f'Cytokit SPRM assay with uuid {self._uuid} has empty view config'
+                raise CytokitSPRMViewConfigError(message)
+            confs.append(conf)
+        return ConfCells(confs, None)
